docker/Sub/requirements.txt
argcomplete==1.12.3
cached-property==1.5.2
paho-mqtt==1.6.1